[DEV-5113] new name space redirectModal ==> Modal
Making this prop in redux more versatille here shortly!
